+++ packages/cli/src/utils/getDirectories.test.js
@@ -26,25 +26,7 @@ test('default directories', () => {
     build: '/test/config/.lowdefy/server/build',
     config: '/test/config',
     dev: '/test/config/.lowdefy/dev',
-    dotLowdefy: '/test/config/.lowdefy',
     server: '/test/config/.lowdefy/server',
-  });
-});
-
-test('specify outputDirectory in options', () => {
-  const directories = getDirectories({
-    configDirectory: '/test/config',
-    options: {
-      outputDirectory: '/test/out',
-    },
-  });
-
-  expect(directories).toEqual({
-    build: '/test/out/server/build',
-    config: '/test/config',
-    dev: '/test/out/dev',
-    dotLowdefy: '/test/out',
-    server: '/test/out/server',
   });
 });
 
@@ -57,10 +39,9 @@ test('specify serverDirectory in options', () => {
   });
 
   expect(directories).toEqual({
-    build: '/test/config/.lowdefy/server/build',
+    build: '/test/server/build',
     config: '/test/config',
     dev: '/test/config/.lowdefy/dev',
-    dotLowdefy: '/test/config/.lowdefy',
     server: '/test/server',
   });
 });
@@ -77,7 +58,6 @@ test('specify devDirectory in options', () => {
     build: '/test/config/.lowdefy/server/build',
     config: '/test/config',
     dev: '/test/dev',
-    dotLowdefy: '/test/config/.lowdefy',
     server: '/test/config/.lowdefy/server',
   });
 });